Use TabulatedFluidProperties base class for bilinear and bicubic
